catalogs/sen1floods11/requirements.txt
click==7.1.2
google-cloud-storage==1.29.0
pystac==0.3.4
rasterio==1.1.5
shapely==1.7.0
